move auto-changing providerUrl
